% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-utils-filter-spct.R
\name{A_internal_label}
\alias{A_internal_label}
\alias{A_total_label}
\title{Absorbance axis labels}
\usage{
A_internal_label(unit.exponent = 0, format = getOption("photobiology.math",
  default = "R.expression"))

A_total_label(unit.exponent = 0, format = getOption("photobiology.math",
  default = "R.expression"))
}
\arguments{
\item{unit.exponent}{integer}

\item{format}{character string, "R", "R.expresion", "R.character", or
"LaTeX".}
}
\value{
a character string or an R expression.
}
\description{
Generate cps axis labels in SI units,
using SI scale factors. Output can be selected as character, expression (R
default devices) or LaTeX (for tikz device).
}
\examples{
A_internal_label()
A_internal_label(3)
A_internal_label(format = "R.expression")
A_internal_label(format = "LaTeX")
A_internal_label(3, format = "LaTeX")

A_total_label()
A_total_label(3)
A_total_label(format = "R.expression")
A_total_label(format = "LaTeX")
A_total_label(3, format = "LaTeX")

}
