% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{apply_normalization}
\alias{apply_normalization}
\title{Apply normalization argument}
\usage{
apply_normalization(x, norm)
}
\arguments{
\item{x}{An R object.}

\item{norm}{numeric or character No longer supported, normalization is always
updated in \code{autoplot()} if present and a unit conversion applied.}
}
\description{
Apply the normalization issuing a message.
}
\keyword{internal}
