% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-chart.R
\name{black_or_white}
\alias{black_or_white}
\title{Chose black vs. white color based on weighted mean of RGB channels}
\usage{
black_or_white(colors, threshold = 0.45)
}
\arguments{
\item{colors}{character A vector of color definitions.}

\item{threshold}{numeric in range 0 to 1.}
}
\description{
Chose black or white color based on a color to be used as background.
Usefull when using \code{geom_text} on top of tiles or bars, or
\code{geom_label} with a variable fill.
}
\examples{

black_or_white("red")
black_or_white(colors()[1:10])

}
