% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-utils.R
\name{check_idfactor_arg}
\alias{check_idfactor_arg}
\title{Check idfactor argument and retrieve if needed}
\usage{
check_idfactor_arg(object, idfactor)
}
\arguments{
\item{object}{generic_spct object.}

\item{idfactor}{character or NULL The name of the factor identifying
spectra when stored in long form.}
}
\description{
Check idfactor argument and retrieve if needed
}
\keyword{internal}
