% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{duration2character}
\alias{duration2character}
\title{Convert lubridate duration objects to a string if possible}
\usage{
duration2character(time.unit)
}
\arguments{
\item{time.unit}{lubridate::duration object or character}
}
\value{
A character string describing a time base or exposure duration.
}
\description{
Convert lubridate duration objects to a string if possible
}
\keyword{internal}
