% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-utils-calibration-spct.R
\name{multipliers_label}
\alias{multipliers_label}
\title{Calibration multipliers axis labels}
\usage{
multipliers_label(
  unit.exponent = 0,
  format = getOption("photobiology.math", default = "R.expression"),
  label.text = axis_labels()[["e.mult"]],
  scaled = FALSE,
  normalized = FALSE
)
}
\arguments{
\item{unit.exponent}{integer}

\item{format}{character string, "R", "R.expresion", "R.character", or
"LaTeX".}

\item{label.text}{character Textual portion of the labels.}

\item{scaled}{logical If \code{TRUE} relative units are assumed.}

\item{normalized}{logical (\code{FALSE}) or numeric Normalization wavelength
in manometers (nm).}
}
\value{
a character string or an R expression.
}
\description{
Calibration multipliers axis labels. Output can be selected
as character, expression (R default devices) or LaTeX (for tikz device).
}
\examples{

multipliers_label()
multipliers_label(3)
multipliers_label(format = "R.expression")
multipliers_label(format = "R.character")
multipliers_label(format = "LaTeX")
multipliers_label(3, format = "LaTeX")

}
