% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{normalization_label}
\alias{normalization_label}
\title{Get a normalization label}
\usage{
normalization_label(spct, digits = 1L)
}
\description{
Handle numeric and non-numeric values, as well as long form spectra
}
\keyword{internal}
