% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{plot.generic_spct}
\alias{plot.generic_spct}
\alias{plot.generic_mspct}
\title{Warn if plot methods is called.}
\usage{
\method{plot}{generic_spct}(x, ...)

\method{plot}{generic_mspct}(x, ...)
}
\value{
an empty \code{ggplot} object.
}
\description{
No automatic plot method is possible for objects of class \code{geenric_spct}
as this class is meant to be used only as a pure base class for derivation.
}
\keyword{internal}
