% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{rename_idfactor}
\alias{rename_idfactor}
\title{Update idfactor name in object}
\usage{
rename_idfactor(object, idfactor)
}
\arguments{
\item{object}{generic_spct object.}

\item{idfactor}{character The name of the factor identifying spectra when
stored in long form.}
}
\description{
Conditionally call \code{\link[photobiology]{setIdFactor}}.
}
\note{
There is no check for \code{NA} or \code{NULL} as we assume that
  \code{\link{check_idfactor_arg}} has already been called on \code{idfactor}
  to check its sanity.
}
\keyword{internal}
