% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.integral.R
\name{stat_integral}
\alias{stat_integral}
\title{Integrate range under curve.}
\usage{
stat_integral(mapping = NULL, data = NULL, geom = "text", range = NULL,
  integral.fun = photobiology::integrate_irradiance, label.fmt = "\%s",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_string}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{range}{a numeric vector of at least length two.}

\item{integral.fun}{function on $x$ and $y$.}

\item{label.fmt}{character string giving a format definition for converting
y-integral values into character strings by means of function \code{\link{sprintf}}.}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}
}
\description{
\code{stat_integral} computes the area under a curve.
}
\section{Computed variables}{

\describe{
  \item{label}{intergral value as formatted text}
  \item{x}{range-midpoint}
  \item{xmin}{range minimum}
  \item{xmax}{range maximum}
  \item{y}{integral value as numeric}
}
}
\examples{
library(photobiology)
library(ggplot2)
ggplot(sun.spct, aes(w.length, s.e.irrad)) + geom_line() +
  stat_integral(geom = "hline")
ggplot(sun.spct, aes(w.length, s.e.irrad)) + geom_line() +
 stat_integral(label.fmt = "\%.3f", color = "red")#' @export

}
\seealso{
Other stats.functions: \code{\link{stat_color}},
  \code{\link{stat_peaks}}
}

