% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-spikes.R
\name{stat_spikes}
\alias{stat_spikes}
\title{Find spikes}
\usage{
stat_spikes(
  mapping = NULL,
  data = NULL,
  geom = "point",
  position = "identity",
  ...,
  height.threshold = 10,
  z.threshold = 9,
  k = 20,
  spike.direction = "both",
  max.spike.width = NA,
  chroma.type = "CMF",
  label.fmt = "\%.3g",
  x.label.fmt = label.fmt,
  y.label.fmt = label.fmt,
  x.label.transform = function(x) {
     x
 },
  y.label.transform = function(x) {
     x
 },
  x.colour.transform = x.label.transform,
  na.rm = FALSE,
  show.legend = FALSE,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs to be set
at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points
on this layer}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This can
include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{height.threshold}{numeric The minimum height of spikes expressed
relative to the median amplitude of the baseline local variation of
\code{x}.}

\item{z.threshold}{numeric Modified local \eqn{Z} values larger than
\code{z.threshold} are detected as boundaries of spikes.}

\item{k}{integer width of median window used for smoothing; must be odd}

\item{spike.direction}{character Controls the direction of spikes to be
detected. Accepted arguments are \code{"up"}, \code{"down"},
\code{"both"}.}

\item{max.spike.width}{integer No longer in use.}

\item{chroma.type}{character one of "CMF" (color matching function) or "CC"
(color coordinates) or a \code{\link[photobiology]{chroma_spct}} object.}

\item{label.fmt, x.label.fmt, y.label.fmt}{character  strings giving a format
definition for construction of character strings labels with function
\code{\link{sprintf}} from \code{x} and/or \code{y} values.}

\item{x.label.transform, y.label.transform, x.colour.transform}{function Applied
to \code{x} or \code{y} values when constructing the character labels or
computing matching colours.}

\item{na.rm}{a logical value indicating whether NA values should be
stripped before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
A data frame of observations found in the data matching the criterion
  of being part of a spike. That is to say, the returned data frame not only
  includes the observation at the tip of the spike but also those on its
  shoulders.
}
\description{
\code{stat_spikes} finds at which \code{x} positions spikes are located.
Spikes can be either upwards or downwards from the baseline.
\strong{Axis flipping is currently not supported.}
}
\details{
This stat uses \code{geom_point} by default as it is the geom most
  likely to work well in almost any situation without need of tweaking. The
  default aesthetics set by this stat allow its direct use with
  \code{geom_text}, \code{geom_label}, \code{geom_line}, \code{geom_rug},
  \code{geom_hline} and \code{geom_vline}. The formatting of the labels
  returned can be controlled by the user.
}
\note{
This stat works nicely together with geoms \code{geom_text_repel} and
  \code{geom_label_repel} from package \code{\link[ggrepel]{ggrepel}} to
  solve the problem of overlapping labels
  by displacing them. To discard overlapping labels use \code{check_overlap =
  TRUE} as argument to \code{geom_text}.

  By default the labels are character values suitable to be plotted as is,
  but with a suitable \code{label.fmt} argument labels suitable for parsing
  by the geoms (e.g., into expressions containing greek letters or super or
  subscripts) can be also easily obtained.
}
\section{Computed variables}{

\describe{
  \item{x}{x-value at the peak (or valley) as numeric}
  \item{y}{y-value at the peak (or valley) as numeric}
  \item{x.label}{x-value of observations in the spike formatted as character}
  \item{y.label}{y-value of observations in the spike formatted as character}
  \item{wl.color}{color definition calculated by assuming that x-values are
  wavelengths expressed in nanometres.}
  \item{BW.color}{color definition that is either "black" or "white", to ensure
  high contrast to \code{wl.color}.}
}
}

\section{Default aesthetics}{

Set by the statistic and available to geoms.
\describe{
  \item{label}{stat(x.label)}
  \item{xintercept}{stat(x)}
  \item{yintercept}{stat(y)}
  \item{fill}{stat(wl.color)}
}
}

\section{Required aesthetics}{

Required by the statistic and need to be set with \code{aes()}.
\describe{
  \item{x}{numeric, wavelength in nanometres}
  \item{y}{numeric, a spectral quantity}
}
}

\section{Spike detection}{


  Spikes are detected based on a modified \eqn{Z} score calculated
  from the differenced spectrum. The \eqn{Z} threshold used should be
  adjusted to the characteristics of the input and desired sensitivity. The
  lower the threshold the more stringent the test becomes, with shorter
  spikes being detected.

  \strong{The algorithms assume a consistent step size for the underlying
  independent variable, e.g., wavelength or time, and should not be applied
  if the data do not fulfil this assumption, at least approximately. As
  \code{find_spkikes()} operates on a single vector, checking this remains
  the responsibility of calling functions or methods such as
  \code{\link[photobiology]{spikes}()} and \code{\link[photobiology]{despike}()}.}

  The algorithm uses running differences to detect abrupt changes in value,
  compared to an estimate of the baseline variation of the differences,
  approximating a baseline \eqn{Z} from MAD and a baseline value from the
  median differences. Currently, a single estimate of MAD is used but running
  medians, when possible, as baseline. This comparison detects running
  differences that are unusually large, in most cases signalling a transition
  between values near the baseline and far from it, in both directions.

  Transitions into- and out of spikes are distinguished based on the median
  of the non-differenced values, as a descriptor of the data baseline. As for
  the median of the differences, a running median is used when possible.

  This function thus detects the start and end of each spike, and
  distinguishes upward and downward spikes.

  \code{k} is the width in number of observations of the window used for
  running median smoothing to extract the baseline. A value several times the
  width of the broader spike but narrow enough to track broader peaks needs
  to be manually set in most cases.

  With \code{na.rm = TRUE}, \code{NA} values are omitted before searching for
  spikes and set to \code{0L} in the returned vector.

  If all spikes are guaranteed to be one observation-wide and either going up
  or down from the baseline, it is possible to detect them based purely on
  the \code{z.threshold} by passing \code{height.threshold = NA} and either
  \code{spike.direction = "up"} or \code{spike.direction = "down"}, which
  ensures very fast computation.

  Parameters of the algorithm need to be adjusted depending on the data, so
  inspection of returned values is needed together with adjustment by trial
  and error of suitable values for \code{z.threshold},
  \code{height.threshold}, and \code{k}.

}

\examples{

# ggplot() methods for spectral objects set a default mapping for x and y.

# two spurious(?) spikes
ggplot(sun.spct) +
  stat_spikes(colour = "red") +
  geom_line()

# no spikes detected
ggplot(sun.spct) +
  stat_spikes(colour = "red", z.threshold = 12) +
  geom_line()

# small noise spikes detected
ggplot(white_led.raw_spct) +
  stat_spikes(colour = "red") +
  geom_line()

ggplot(white_led.raw_spct) +
  stat_spikes(colour = "red") +
  stat_spikes(geom = "text", colour = "red", check_overlap = TRUE,
             vjust = -0.5, label.fmt = "\%3.0f nm") +
  geom_line()

# noisy data
ggplot(white_led.raw_spct, aes(w.length, counts_2)) +
  stat_spikes(colour = "red",
              z.threshold = 10) +
  geom_line()

# with height.threshold to ignore minor spikes
ggplot(white_led.raw_spct, aes(w.length, counts_2)) +
  stat_spikes(colour = "red", height.threshold = 15, z.threshold = 10) +
  geom_line()

# colour from wavelengths (here black for UV and NIR)
ggplot(white_led.raw_spct, aes(w.length, counts_2)) +
  stat_spikes(mapping = aes(colour = after_stat(wl.color)),
              height.threshold = 15, z.threshold = 10) +
  geom_line() +
  scale_colour_identity()

}
\seealso{
\code{\link[photobiology]{find_spikes}}, which is used internally.

Other stats functions:
\code{\link[=stat_color]{stat_color()}},
\code{\link[=stat_find_qtys]{stat_find_qtys()}},
\code{\link[=stat_find_wls]{stat_find_wls()}},
\code{\link[=stat_label_peaks]{stat_label_peaks()}},
\code{\link[=stat_peaks]{stat_peaks()}},
\code{\link[=stat_wb_box]{stat_wb_box()}},
\code{\link[=stat_wb_column]{stat_wb_column()}},
\code{\link[=stat_wb_contribution]{stat_wb_contribution()}},
\code{\link[=stat_wb_hbar]{stat_wb_hbar()}},
\code{\link[=stat_wb_irrad]{stat_wb_irrad()}},
\code{\link[=stat_wb_label]{stat_wb_label()}},
\code{\link[=stat_wb_mean]{stat_wb_mean()}},
\code{\link[=stat_wb_relative]{stat_wb_relative()}},
\code{\link[=stat_wb_sirrad]{stat_wb_sirrad()}},
\code{\link[=stat_wb_total]{stat_wb_total()}},
\code{\link[=stat_wl_strip]{stat_wl_strip()}},
\code{\link[=stat_wl_summary]{stat_wl_summary()}}
}
\concept{stats functions}
