% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat-wb-contribution.R
\name{stat_wb_contribution}
\alias{stat_wb_contribution}
\title{Integrate ranges under spectral curve.}
\usage{
stat_wb_contribution(
  mapping = NULL,
  data = NULL,
  geom = "text",
  position = "identity",
  ...,
  w.band = NULL,
  integral.fun = integrate_xy,
  label.mult = 1,
  chroma.type = "CMF",
  label.fmt = "\%1.2f",
  ypos.mult = 1.07,
  ypos.fixed = NULL,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{The aesthetic mapping, usually constructed with
\code{\link[ggplot2]{aes}} or \code{\link[ggplot2]{aes_}}. Only needs
to be set at the layer level if you are overriding the plot defaults.}

\item{data}{A layer specific dataset - only needed if you want to override
the plot defaults.}

\item{geom}{The geometric object to use display the data}

\item{position}{The position adjustment to use for overlapping points on this
layer}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. This
can include aesthetics whose values you want to set, not map. See
\code{\link[ggplot2]{layer}} for more details.}

\item{w.band}{a waveband object or a list of waveband objects or numeric
vector of at least length two.}

\item{integral.fun}{function on $x$ and $y$.}

\item{label.mult}{numeric Scaling factor applied to y-integral values before
conversion into character strings.}

\item{chroma.type}{character one of "CMF" (color matching function) or "CC"
(color coordinates) or a \code{\link[photobiology]{chroma_spct}} object.}

\item{label.fmt}{character string giving a format definition for converting
y-integral values into character strings by means of function
\code{\link{sprintf}}.}

\item{ypos.mult}{numeric Multiplier constant used to scale returned
\code{y} values.}

\item{ypos.fixed}{numeric If not \code{NULL} used a constant value returned
in \code{y}.}

\item{na.rm}{a logical value indicating whether NA values should be stripped
before the computation proceeds.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics, rather
than combining with them. This is most useful for helper functions that
define both data and aesthetics and shouldn't inherit behaviour from the
default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\value{
A data frame with one row for each waveband object in the argument
to \code{w.band}. Wavebeand outside the range of the spectral data are
trimmed or discarded.
}
\description{
\code{stat_wb_contribution} integrates the area under a spectral curve. It
first integrates the area under the curve for each waveband and for the whole
curve and then expresses the integral for each band as a relative
contribution to the area under the whole spectral curve. Sets suitable
default aesthetics for "rect", "hline", "vline", "text" and "label" geoms
displaying "contributions" per waveband to the total of the spectral
integral. \strong{\code{x}-scale transformations and axis flipping are
currently not supported}.
}
\section{Computed variables}{

What it is named integral below is the result of appying \code{integral.fun}
to the data, with default \code{integrate_xy}.
\describe{
  \item{y.label}{yint multiplied by \code{label.mult} and formatted
  according to \code{label.fmt}}
  \item{x}{w.band-midpoint}
  \item{xmin}{w.band minimum}
  \item{xmax}{w.band maximum}
  \item{ymin}{data$y minimum}
  \item{ymax}{data$y maximum}
  \item{yint}{data$y integral for w.band / data$y integral for whole range
  of data$x}
  \item{xmean}{yint divided by wl_expanse(w.band)}
  \item{y}{ypos.fixed or top of data, adjusted by \code{ypos.mult}}
  \item{wb.color}{color of the w.band}
  \item{wb.name}{label of w.band}
}
}

\section{Default aesthetics}{

Set by the statistic and available to geoms.
\describe{
  \item{label}{..y.label..}
  \item{x}{..x..}
  \item{xmin}{..xmin..}
  \item{xmax}{..xmax..}
  \item{ymin}{..y.. - (..ymax.. - ..ymin..) * 0.03}
  \item{ymax}{..y.. + (..ymax.. - ..ymin..) * 0.03}
  \item{yintercept}{..ymean..}
  \item{fill}{..wb.color..}
}
}

\section{Required aesthetics}{

Required by the statistic and need to be set with \code{aes()}.
\describe{
  \item{x}{numeric, wavelength in nanometres}
  \item{y}{numeric, a spectral quantity}
}
}

\examples{

library(photobiologyWavebands)
# ggplot() methods for spectral objects set a default mapping for x and y.

# Using defaults
ggplot(sun.spct) +
  geom_line() +
  stat_wb_box(w.band = VIS()) +
  stat_wb_contribution(w.band = VIS()) +
  scale_fill_identity() + scale_color_identity()

# Setting position and angle of the text
ggplot(sun.spct) +
  geom_line() +
  stat_wb_box(w.band = VIS_bands()) +
  stat_wb_contribution(w.band = VIS_bands(), angle = 90, size = 2.5) +
  scale_fill_identity() + scale_color_identity()

# Showing percentages, i.e., using a different format for numbers
ggplot(sun.spct) +
  geom_line() +
  stat_wb_box(w.band = VIS_bands()) +
  stat_wb_contribution(w.band = VIS_bands(), size = 2.5,
                       label.mult = 100, label.fmt = "\%3.0f\%\%") +
  scale_fill_identity() + scale_color_identity()

# Including the name of the waveband, i.e., changing the mapping for label
ggplot(sun.spct, range = c(NA, 410)) +
  geom_line() +
  stat_wb_box(w.band = UV_bands(), color = "white") +
  stat_wb_contribution(w.band = UV_bands(), size = 2.5,
                       label.mult = 100, label.fmt = "\%3.0f\%\%",
                       mapping = aes(label = after_stat(paste(wb.name, y.label)))) +
  scale_fill_identity() + scale_color_identity()

}
\seealso{
Other stats functions:
\code{\link[=stat_color]{stat_color()}},
\code{\link[=stat_find_qtys]{stat_find_qtys()}},
\code{\link[=stat_find_wls]{stat_find_wls()}},
\code{\link[=stat_label_peaks]{stat_label_peaks()}},
\code{\link[=stat_peaks]{stat_peaks()}},
\code{\link[=stat_spikes]{stat_spikes()}},
\code{\link[=stat_wb_box]{stat_wb_box()}},
\code{\link[=stat_wb_column]{stat_wb_column()}},
\code{\link[=stat_wb_hbar]{stat_wb_hbar()}},
\code{\link[=stat_wb_irrad]{stat_wb_irrad()}},
\code{\link[=stat_wb_label]{stat_wb_label()}},
\code{\link[=stat_wb_mean]{stat_wb_mean()}},
\code{\link[=stat_wb_relative]{stat_wb_relative()}},
\code{\link[=stat_wb_sirrad]{stat_wb_sirrad()}},
\code{\link[=stat_wb_total]{stat_wb_total()}},
\code{\link[=stat_wl_strip]{stat_wl_strip()}},
\code{\link[=stat_wl_summary]{stat_wl_summary()}}
}
\concept{stats functions}
