% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-plot-title.R
\name{subtitle_spct.generic_spct}
\alias{subtitle_spct.generic_spct}
\title{Build a plot subtitle from object metadata (deprecated!)}
\usage{
subtitle_spct.generic_spct(
  x,
  default.when = character(0),
  default.where = character(0),
  ...
)
}
\arguments{
\item{x}{An R object.}

\item{default.when}{datetime object.}

\item{default.where}{One-row data.frame with at least variables \code{lat},
and \code{lon}.}

\item{...}{Additional parameters of derived methods.}
}
\description{
Build a character string to be used as plot title by quering spectral
object for its  metadata.
}
\keyword{internal}
