% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-plot-title.R
\name{title_spct}
\alias{title_spct}
\title{Build a plot title from object metadata (deprecated!)}
\usage{
title_spct(x, default.what = character(0), ...)
}
\arguments{
\item{x}{An R object.}

\item{default.what}{A character string.}

\item{...}{Additional parameters of derived methods.}
}
\description{
Build a character string to be used as plot title by quering spectral
object for its  metadata.
}
\keyword{internal}
