% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-utils.R
\name{validate_geom_arg}
\alias{validate_geom_arg}
\title{Validate argument passed to geom}
\usage{
validate_geom_arg(
  geoms = character(),
  supported.geoms = c("area", "line", "spct", "col", "point"),
  default.geoms = "line"
)
}
\arguments{
\item{geoms, supported.geoms, default.geoms}{character A vector of geom names.}
}
\description{
\code{autoplot()} methods have a geom parameter that supports a subset of
existing geoms. This function factors this check out of the methods' code.
}
\keyword{internal}
