% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-utils.R
\name{validate_idfactor}
\alias{validate_idfactor}
\title{Validate idfactor}
\usage{
validate_idfactor(idfactor)
}
\arguments{
\item{idfactor}{A character string, NULL or NA}
}
\description{
Validate idfactor
}
\keyword{internal}
