% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-utils-wlength.R
\name{w_number}
\alias{w_number}
\alias{w_frequency}
\title{Deprecated functions}
\usage{
w_number(w.length, unit.exponent = 0)

w_frequency(w.length, unit.exponent = 0)
}
\arguments{
\item{w.length}{numeric wavelength (nm)}

\item{unit.exponent}{integer Exponent of the scale multiplier implicit in
result, e.g., use 3 for kJ.}
}
\description{
To convert wavelength into wavenumber or into frequency, please, use the
conversion functions from package 'photobiology' in place of the deprecated
functions \code{w_number()} and \code{w_frequency()} from this package.
}
\section{Deprecated}{
 These functions will be removed from package 'ggpmisc'
  in the near future.
}

\examples{
library(photobiology)

wl2wavenumber(600)
wl2frequency(600)

}
\seealso{
See \code{\link[photobiology]{wl2wavenumber}} for the functions to
  be used in all new code.
}
