% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{warn_norm_arg}
\alias{warn_norm_arg}
\title{Warn if norm argument passed in call}
\usage{
warn_norm_arg(norm)
}
\arguments{
\item{norm}{numeric or character No longer supported, normalization is always
updated in \code{autoplot()} if present and a unit conversion applied.}
}
\description{
Warn if norm argument passed in call
}
\keyword{internal}
