% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{autoplot.generic_spct}
\alias{autoplot.generic_spct}
\alias{autoplot.generic_mspct}
\title{Warn if plot methods is called.}
\usage{
\method{autoplot}{generic_spct}(object, ...)

\method{autoplot}{generic_mspct}(object, ...)
}
\value{
an empty \code{ggplot} object.
}
\description{
No automatic plot method is possible for objects of class \code{geenric_spct}
as this class is meant to be used only as a pure base class for derivation.
}
\keyword{internal}
