% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-waveband.R
\name{autoplot.waveband}
\alias{autoplot.waveband}
\title{Create a complete ggplot for a waveband descriptor.}
\usage{
\method{autoplot}{waveband}(
  object,
  ...,
  w.length = NULL,
  range = NULL,
  fill = 0,
  span = NULL,
  wls.target = "HM",
  unit.in = getOption("photobiology.radiation.unit", default = "energy"),
  unit.out = unit.in,
  annotations = NULL,
  by.group = FALSE,
  geom = "line",
  wb.trim = TRUE,
  norm = NA,
  text.size = 2.5,
  ylim = c(NA, NA),
  object.label = deparse(substitute(object)),
  na.rm = TRUE
)
}
\arguments{
\item{object}{a waveband object.}

\item{...}{arguments passed along by name to \code{autoplot.response_spct()}.}

\item{w.length}{numeric vector of wavelengths (nm).}

\item{range}{an R object on which range() returns a vector of length 2, with
min annd max wavelengths (nm).}

\item{fill}{value to use as response for wavelengths outside the waveband
range.}

\item{span}{a peak is defined as an element in a sequence which is greater
than all other elements within a window of width span centered at that
element.}

\item{wls.target}{numeric vector indicating the spectral quantity values for
which wavelengths are to be searched and interpolated if need. The
\code{character} strings "half.maximum" and "half.range" are also accepted
as arguments. A list with \code{numeric} and/or \code{character} values is
also accepted.}

\item{unit.in, unit.out}{the type of unit we assume as reference: "energy" or
"photon" based for the waveband definition and the implicit matching
response plotted.}

\item{annotations}{a character vector. For details please see section Plot
Annotations.}

\item{by.group}{logical flag If TRUE repeated identical annotation layers are
added for each group within a plot panel as needed for animation. If
\code{FALSE}, the default, single layers are added per panel.}

\item{geom}{character The name of a ggplot geometry, currently only
\code{"area"}, \code{"spct"} and \code{"line"}.}

\item{wb.trim}{logical. Passed to \code{\link[photobiology]{trim_wl}}.
Relevant only when the \code{waveband} extends partly outside \code{range}.}

\item{norm}{numeric or character Normalization wavelength (nm) or character
string \code{"max"} or other criterion for normalization.}

\item{text.size}{numeric size of text in the plot decorations.}

\item{ylim}{numeric y axis limits,}

\item{object.label}{character The name of the object being plotted.}

\item{na.rm}{logical.}
}
\value{
a \code{ggplot} object.
}
\description{
Construct a ggplot object with an annotated plot of a \code{waveband} object.
}
\details{
A \code{response_spct} object is created based on the
  \code{waveband} object and the argument passed to parameter \code{w.length}.
  By default wavelengths spanning the waveband definition expanded by
  1 nm at each end are used. A \code{waveband} object can describe either a
  simple wavelength range or a (biological) spectral weighting function
  (BSWF). An effectiveness is a response expressed per unit of excitation,
  and in most cases normalised.

  Effectiveness spectra can be plotted expressing the spectral effectiveness
  either per mol of photons (\eqn{1 mol^{-1} nm}) or per joule of energy
  (\eqn{1 J^{-1} nm}), selected through formal argument \code{unit.out}. The
  value of \code{unit.in} has no direct effect on the result for BSWFs,
  as BSWFs are defined based on a certain base of expression, which is
  enforced. Indirectly it affects plots as \code{unit.out} defaults to
  \code{unit.in}. In contrast, for wavebands which only define a wavelength
  range, changing the assumed reference irradiance units, changes the
  responsivity according to Plank's law, i.e., the four possible combinations
  of pairs of values for \code{unit.in} and \code{unit.out} produce four
  different plots. _Only in rare cases \code{unit.in} and \code{unit.out}
  is useful as normally the dependency on base of expresion is encoded in
  the \code{waveband} definition as a BSWF._

  While \code{w.length}` provides the wavelengths of the generated
  response spectrum, `range` sets the wavelength range of the \eqn{x}-axis
  of the plot.

  Unused named arguments are forwarded to
  \code{\link{autoplot.response_spct}()}, allowing control of additional plot
  properties.
}
\section{Plot Annotations}{
 The recognized annotation names are: \code{"summaries"}, \code{"peaks"},
  \code{"peak.labels"}, \code{"valleys"}, \code{"valley.labels"},
  \code{"wls"}, \code{"wls.labels"}, \code{"colour.guide"},
  \code{"color.guide"}, \code{"boxes"}, \code{"segments"}, \code{"labels"}.
  In addition, \code{"+"} is interpreted as a request to add to the already
  present default annotations, \code{"-"} as request to remove annotations
  and \code{"="} or missing\code{"+"} and \code{"-"} as a request to reset
  annotations to those requested. If used, \code{"+"}, \code{"-"} or
  \code{"="} must be the first member of a character vector, and followed by
  one or more of the names given above. To simultaneously add and remove
  annotations one can pass a \code{list} containing \code{character} vectors
  each assembled as described. The vectors are applied in the order they
  appear in the list. To disable all annotations pass \code{""} or
  \code{c("=", "")} as argument. Adding a variation of an annotation already
  present, replaces the existing one automatically: e.g., adding
  \code{"peak.labels"} replaces\code{"peaks"} if present.

  The annotation layers are added to the plot using statistics defined in 'ggspectra':
  \code{\link{stat_peaks}}, \code{\link{stat_valleys}},
  \code{\link{stat_label_peaks}}, \code{\link{stat_label_valleys}},
  \code{\link{stat_find_wls}}, \code{\link{stat_spikes}},
  \code{\link{stat_wb_total}}, \code{\link{stat_wb_mean}},
  \code{\link{stat_wb_irrad}}, \code{\link{stat_wb_sirrad}},
  \code{\link{stat_wb_contribution}}, \code{\link{stat_wb_relative}},
  and \code{\link{stat_wl_strip}}. However, only some of their parameters
  can be passed arguments through \code{autoplot} methods. In some cases
  the defaults used by \code{autoplot} methods are not the defaults of the
  statistics.
}

\examples{

autoplot(waveband(c(400, 500)))
autoplot(waveband(c(400, 500)), w.length = c(300,600))
autoplot(waveband(c(400, 500)), range = c(300,600))
autoplot(waveband(c(400, 500)), geom = "spct")

}
\seealso{
\code{\link{autoplot.response_spct}},
  \code{\link[photobiology]{waveband}}.

Other autoplot methods: 
\code{\link{autoplot.calibration_spct}()},
\code{\link{autoplot.cps_spct}()},
\code{\link{autoplot.filter_spct}()},
\code{\link{autoplot.object_spct}()},
\code{\link{autoplot.raw_spct}()},
\code{\link{autoplot.reflector_spct}()},
\code{\link{autoplot.response_spct}()},
\code{\link{autoplot.source_spct}()}
}
\concept{autoplot methods}
