% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-labels-texts.R
\name{axis_labels}
\alias{axis_labels}
\alias{axis_labels_uk}
\alias{axis_labels_uk_comma}
\alias{axis_labels_none}
\title{Default text for axis labels}
\usage{
axis_labels()

axis_labels_uk()

axis_labels_uk_comma()

axis_labels_none()
}
\value{
A character vector
}
\description{
Obtain texts used by default for axis labels in
  plots. They contain only the text part, but not symbols or units of
  expression. Can be used to change the language or to suppress the text.
}
\examples{
axis_labels()[["w.length"]]

}
