% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{check_idfactor_arg}
\alias{check_idfactor_arg}
\title{Check idfactor argument and retrieve if needed}
\usage{
check_idfactor_arg(object, idfactor = NULL, default = FALSE)
}
\arguments{
\item{object}{generic_spct object.}

\item{idfactor}{character, NULL or NA The name of the factor identifying
spectra when stored in long form.}

\item{default}{logical}
}
\description{
Check idfactor argument and retrieve if needed
}
\keyword{internal}
