% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-utils-raw-spct.R
\name{counts_label}
\alias{counts_label}
\title{Raw-counts axis labels}
\usage{
counts_label(unit.exponent = 3, format = getOption("photobiology.math",
  default = "R.expression"))
}
\arguments{
\item{unit.exponent}{integer}

\item{format}{character string, "R", "R.expresion", "R.character", or
"LaTeX".}
}
\value{
a character string or an R expression.
}
\description{
Generate axis labels in SI units,
using SI scale factors. Output can be selected as character, expression (R
default devices) or LaTeX (for tikz device).
}
\examples{

counts_label()
counts_label("R.expression")
counts_label("LaTeX")

}
