% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decode_annotations}
\alias{decode_annotations}
\title{Merge user supplied annotations with default ones}
\usage{
decode_annotations(annotations, annotations.default = "colour.guide")
}
\arguments{
\item{annotations, annotations.default}{character vector or a list of
character vectors.}
}
\description{
Allow users to add and subract from default annotations in addition
to providing a given set of annotations.
}
\keyword{internal}
