% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoration.R
\name{decoration}
\alias{decoration}
\title{Add decorations to spectrum plot (private)}
\usage{
decoration(
  w.band,
  y.max,
  y.min,
  x.max,
  x.min,
  annotations,
  span = NULL,
  strict = is.null(span),
  wls.target = "HM",
  label.qty,
  label.mult = 1,
  summary.label,
  unit.out = NULL,
  time.unit = NULL,
  text.size = 2.5,
  label.color = NULL,
  chroma.type = "CMF",
  pos.shift = 0,
  na.rm = TRUE
)
}
\arguments{
\item{w.band}{waveband object or list of waveband objects}

\item{annotations}{character vector}

\item{span}{numeric}

\item{strict}{logical}

\item{wls.target}{numeric or character vector}

\item{label.qty}{character}

\item{summary.label}{character}

\item{text.size}{numeric}

\item{label.color}{color definition or name}

\item{chroma.type}{character one of "CMF" (color matching function) or "CC"
(color coordinates) or a \code{\link[photobiology]{chroma_spct}} object.}

\item{pos.shift}{numeric}

\item{na.rm}{logical}

\item{ymax, ymin, xmax, xmin}{numeric}
}
\value{
A list of ggplot "components" that can be added to a ggplot object
  with operator "+". The length of the list depends on the value of argument
  \code{annotations}.
}
\description{
Add decorations to plots generated by the plot() methods defined in this
package. It collects code that is common to plot methods for different types
of spectra but as it may change in the future it is not exported.
}
\keyword{internal}
