% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-source-spct.r,
%   R/axis-utils-source-spct.R
\name{duration2character}
\alias{duration2character}
\title{Convert lubridate duration objects to a string if possible}
\usage{
duration2character(time.unit)

duration2character(time.unit)
}
\arguments{
\item{time.unit}{lubridate::duration object or character}
}
\description{
Convert lubridate duration objects to a string if possible

Convert lubridate duration objects to a string if possible
}
\keyword{internal}
