% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-spct.r, R/stat-color.R,
%   R/stat-find-wls.R, R/stat-label-peaks.R, R/stat-peaks.R, R/stat-spikes.R,
%   R/stat-wb-box.R, R/stat-wb-column.R, R/stat-wb-contribution.R,
%   R/stat-wb-hbar.R, R/stat-wb-irrad.R, R/stat-wb-label.R, R/stat-wb-mean.R,
%   R/stat-wb-relative.R, R/stat-wb-sirrad.R, R/stat-wb-total.R,
%   R/stat-wl-strip.R, R/stat-wl-summary.R
\docType{data}
\name{GeomSpct}
\alias{GeomSpct}
\alias{StatColor}
\alias{StatFindWls}
\alias{StatFindQty}
\alias{Stats}
\alias{StatLabelPeaks}
\alias{StatLabelValleys}
\alias{StatPeaks}
\alias{StatValleys}
\alias{StatSpikes}
\alias{StatWbBox}
\alias{wb_col_compute_group}
\alias{StatWbColumn}
\alias{StatWbContrib}
\alias{StatWbHbar}
\alias{StatWbIrrad}
\alias{StatWbLabel}
\alias{StatWbMean}
\alias{StatWbRelative}
\alias{StatWbSIrrad}
\alias{StatWbTotal}
\alias{StatColorGuide}
\alias{StatWlSummary}
\title{\code{Stat*} Objects}
\format{
An object of class \code{StatFindWls} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatFindQty} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatLabelValleys} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.

An object of class \code{StatValleys} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
StatFindWls

StatFindQty

StatLabelValleys

StatValleys
}
\description{
All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.

All \code{stat_*} functions (like \code{stat_bin}) return a layer that
contains a \code{Stat*} object (like \code{StatBin}). The \code{Stat*}
object is responsible for rendering the data in the plot.
}
\details{
Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object,
descended from the top-level \code{Stat}, and each implements various methods
and fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:

Each of the \code{Stat*} objects is a \code{\link[ggplot2]{ggproto}} object, descended
from the top-level \code{Stat}, and each implements various methods and
fields. To create a new type of Stat object, you typically will want to
implement one or more of the following:
}
\seealso{
\code{\link[ggplot2]{ggplot2-ggproto}}

\code{\link[ggplot2]{ggplot2-ggproto}}

\code{\link[ggplot2]{ggplot2-ggproto}}

\code{\link[ggplot2]{ggplot2-ggproto}}
}
\keyword{datasets}
\keyword{internal}
