% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-internal.R
\name{rename_idfactor}
\alias{rename_idfactor}
\title{Update idfactor name in object}
\usage{
rename_idfactor(object, idfactor)
}
\arguments{
\item{object}{generic_spct object.}

\item{idfactor}{character The name of the factor identifying spectra when
stored in long form.}
}
\description{
Update idfactor name in object
}
\keyword{internal}
