% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot-utils.R
\name{validate_idfactor}
\alias{validate_idfactor}
\title{Validate idfactor}
\usage{
validate_idfactor(idfactor)
}
\arguments{
\item{idfactor}{character, logical, NULL or NA. If a \code{character} string,
it is validated as an R name, while logical values, NULL and NA are
returned unchanged.}
}
\description{
Validate idfactor
}
\keyword{internal}
