% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis-utils-wlength.R
\name{w_number}
\alias{w_number}
\alias{w_frequency}
\title{Wavelength conversions}
\usage{
w_number(w.length, unit.exponent = 0)

w_frequency(w.length, unit.exponent = 0)
}
\arguments{
\item{w.length}{numeric wavelength (nm)}

\item{unit.exponent}{integer}
}
\description{
Convert wavelength into wavenumber or into frequency.
}
\examples{

w_number(600)
w_frequency(600)

}
